+++ src/system/FPS.cpp
@@ -65,7 +65,7 @@ const std::string FPS::asString ( void ) const
 
 uint32 FPS::fps ( void ) const
 {
-  return this->frames() / ( this->fps_timer.ticks() / 1000.f );
+  return this->frames() / static_cast<uint32> ( ( this->fps_timer.ticks() / 1000.f ) );
 }
 
 void FPS::update ( void )
